Moved enumCaseSensitive property to LoaderOptions and deleted ConstructorOptions class
